ADD: Increment test coverage
