+++ westack/model/cursor.go
@@ -7,7 +7,6 @@ import (
 
 type Cursor interface {
 	Next() (result *Instance, err error)
-	HasNext() bool
 	All() (result InstanceA, err error)
 }
 
@@ -23,10 +22,6 @@ func (cursor *mongoCursor) Next() (result *Instance, err error) {
 	} else {
 		return result, nil
 	}
-}
-
-func (cursor *mongoCursor) HasNext() bool {
-	return cursor.mCursor.Next(cursor.ctx)
 }
 
 func (cursor *mongoCursor) All() (result InstanceA, err error) {
@@ -53,10 +48,6 @@ func (cursor *errorCursor) Next() (result *Instance, err error) {
 	return result, cursor.err
 }
 
-func (cursor *errorCursor) HasNext() bool {
-	return false
-}
-
 func (cursor *errorCursor) All() (result InstanceA, err error) {
 	return result, cursor.err
 }
@@ -81,10 +72,6 @@ func (cursor *fixedLengthCursor) Next() (result *Instance, err error) {
 	return
 }
 
-func (cursor *fixedLengthCursor) HasNext() bool {
-	return cursor.index < len(cursor.instances)
-}
-
 func (cursor *fixedLengthCursor) All() (result InstanceA, err error) {
 	return cursor.instances, nil
 }
@@ -93,54 +80,5 @@ func newFixedLengthCursor(instances InstanceA) Cursor {
 	return &fixedLengthCursor{
 		instances: instances,
 		index:     0,
-	}
-}
-
-type ChannelCursor struct {
-	channel chan *Instance
-	Err     error
-}
-
-func (cursor *ChannelCursor) Next() (result *Instance, err error) {
-	if cursor.Err != nil {
-		return result, cursor.Err
-	}
-	result = <-cursor.channel
-	if result == nil {
-		return result, nil
-	}
-	return
-}
-
-func (cursor *ChannelCursor) HasNext() bool {
-	return cursor.Err == nil
-}
-
-func (cursor *ChannelCursor) All() (result InstanceA, err error) {
-	for {
-		if cursor.Err != nil {
-			return result, cursor.Err
-		}
-		instance := <-cursor.channel
-		if instance == nil {
-			break
-		}
-		result = append(result, *instance)
-	}
-	return
-}
-
-func (cursor *ChannelCursor) Close() error {
-	close(cursor.channel)
-	return nil
-}
-
-func (cursor *ChannelCursor) Error(err error) {
-	cursor.Err = err
-}
-
-func newChannelCursor(channel chan *Instance) Cursor {
-	return &ChannelCursor{
-		channel: channel,
 	}
 }
